payment service was integrated into the order service

+++ inventory-service/src/main/java/com/fardin/inventroyservice/repository/ProductEntryRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.stereotype.Repository;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
@@ -15,6 +16,6 @@ import java.util.Optional;
 public interface ProductEntryRepository extends JpaRepository<ProductEntry, String> {
 
     @Query("SELECT sum(p.quantity) FROM ProductEntry p WHERE p.productId = :productId")
-    public Optional<BigDecimal> calculateInventoryByProductId(@Param("productId") Integer productId);
+    public Optional<BigInteger> calculateInventoryByProductId(@Param("productId") Integer productId);
 
 }
